[server] improved seeding (#72)
* basic, theoretical implementation

* Fix bugs from previous commit, now works in theory, just very slow

* fix wrong seed (hashed version) saved in replay file, xxhash header recompile each time since it was downloaded & overwritten on each compile

* Update server/src/utils/random_string.cpp to use one random device


---------

+++ server/inc/config/worldgen/EmptyWorldGenerator.h
@@ -8,7 +8,7 @@ class EmptyWorldGenerator : public WorldGenerator
 	public:
 		EmptyWorldGenerator();
 
-		void generateWorld(unsigned int seed);
+		void generateWorld(uint64_t seed);
 };
 
 #endif // EMPTY_WORLD_GENERATOR_H
